requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
backoff>=2.0.0
tiktoken>=0.5.0
requests>=2.31.0
pandas>=2.0.0
tqdm>=4.65.0
